added asserts to unit tests
